Move the parser, too
